int32_t to long long int ... check it

+++ CMSSW/src/L1Trigger/TrackFindingAM/interface/PCATrackFitter.h
@@ -43,10 +43,10 @@ class PCATrackFitter:public TrackFitter
   bool useinteger_ ;
 
   std::vector<pca::matrixpcaconst<double> > pcacontvct_float_;
-  std::vector<pca::matrixpcaconst<int32_t> > pcacontvct_integer_;
+  std::vector<pca::matrixpcaconst<long long int> > pcacontvct_integer_;
 
   std::vector<double> chi2vf_;
-  std::vector<int32_t> chi2vi_;
+  std::vector<long long int> chi2vi_;
 
   /* we will merge these using TypeIs... structs */
   void fit_float(vector<Hit*> hits);
@@ -75,7 +75,7 @@ class PCATrackFitter:public TrackFitter
   {
     return chi2vf_;
   }
-  const std::vector<int32_t> & get_chi2i() const
+  const std::vector<long long int> & get_chi2i() const
   {
     return chi2vi_;
   }
